requirements.txt
python-telegram-bot==21.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiohttp==3.9.5
Pillow==10.4.0
playwright==1.46.0
openai==1.43.0
tenacity==8.5.0
mutagen==1.47.0
